Separate some code
