+++ Modules/DebugModule.cs
@@ -1,5 +1,6 @@
 ﻿using Discord;
 using Discord.Commands;
+using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -15,7 +16,7 @@ namespace Assembly_Bot.Modules
         [Summary("Get info about a role.")]
         [RequireOwner]
         [RequireBotPermission(ChannelPermission.SendMessages)]
-        public async Task mentionRoleAsync(IRole role)
+        public async Task MentionRoleAsync(IRole role)
         {
             var embed = new EmbedBuilder();
             embed.WithDescription("Role ID : " + role.Id + Environment.NewLine + "Role Name : " + role.Name + Environment.NewLine + "Role Mention : " + role.Mention + Environment.NewLine + "Role Mention : " + role.Mention + Environment.NewLine + "Role Color : " + role.Color.ToString() + Environment.NewLine + "Role Created at : " + role.CreatedAt);
